+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Improve chunk_query() function
 * Rename queryes -> queries
 * Add debug_info() function
+* Refactor multiquery() decorator
+* Improve BaseDumper.mode_action()
 
 ## 0.1.0.dev3